Safari-609.1.11.1/LayoutTests/accessibility/mac/native-list-item-should-not-expose-label-expected.txt

Some text. Like
This tests that list items on MacOS do not expose labeling information. MacOS expects only leaf-node type elements to do that.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS listitem.description is 'AXDescription: '
PASS listitem.title is 'AXTitle: '
PASS successfullyParsed is true

TEST COMPLETE
